requirements.txt
django
django-channels
